+++ src/main/java/snw/rfm/api/events/GameStopEvent.java
@@ -13,9 +13,7 @@ package snw.rfm.api.events;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
-import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.Set;